Rename 'creature' faker to 'creatures'
Keep plurality consistent across fakers' names
