+++ docs/src/orchid/resources/pages/fakers/creatures.md
@@ -1,9 +1,9 @@
 ## `CreaturesFaker`
 
-[![Maven Central](https://img.shields.io/maven-central/v/io.github.serpro69/kotlin-faker-creature?style=for-the-badge)](https://search.maven.org/artifact/io.github.serpro69/kotlin-faker-creature)
-[![Sonatype Nexus (Snapshots)](https://img.shields.io/nexus/s/io.github.serpro69/kotlin-faker-creature?label=snapshot-version&server=https%3A%2F%2Foss.sonatype.org&style=for-the-badge&color=yellow)](#downloading)
+[![Maven Central](https://img.shields.io/maven-central/v/io.github.serpro69/kotlin-faker-creatures?style=for-the-badge)](https://search.maven.org/artifact/io.github.serpro69/kotlin-faker-creatures)
+[![Sonatype Nexus (Snapshots)](https://img.shields.io/nexus/s/io.github.serpro69/kotlin-faker-creatures?label=snapshot-version&server=https%3A%2F%2Foss.sonatype.org&style=for-the-badge&color=yellow)](#downloading)
 
-Provides access to fake data generators within the Creature domain.
+Provides access to fake data generators within the Creatures domain.
 
 ## Usage
 
@@ -18,7 +18,7 @@ Latest releases are always available on maven central.
 ```groovy
 dependencies {
     implementation 'io.github.serpro69:kotlin-faker:$version'
-    implementation 'io.github.serpro69:kotlin-faker-creature:$version'
+    implementation 'io.github.serpro69:kotlin-faker-creatures:$version'
 }
 ```  
 
@@ -33,19 +33,19 @@ dependencies {
     </dependency>
     <dependency>
         <groupId>io.github.serpro69</groupId>
-        <artifactId>kotlin-faker-creature</artifactId>
+        <artifactId>kotlin-faker-creatures</artifactId>
         <version>${version}</version>
     </dependency>
 </dependencies>
 ```  
 
-_NB! An additional fake data provider like 'creature' requires the main `kotlin-faker` dependency to be on the classpath._
+_NB! An additional fake data provider like 'creatures' requires the main `kotlin-faker` dependency to be on the classpath._
 
 ### Generating data
 
 ```kotlin
 // NB! the package you import if using multiple fakers
-import io.github.serpro69.kfaker.creature.faker
+import io.github.serpro69.kfaker.creatures.faker
 
 val faker = faker { }
