+++ faker/creatures/src/integration/kotlin/io/github/serpro69/kfaker/creatures/CreatureFakerIT.kt
@@ -1,11 +1,11 @@
-package io.github.serpro69.kfaker.creature
+package io.github.serpro69.kfaker.creatures
 
 import io.github.serpro69.kfaker.test.helper.`every public function in each provider is invoked without exceptions`
 import io.github.serpro69.kfaker.test.helper.`faker instance is initialized with custom locale`
 import io.kotest.core.spec.style.DescribeSpec
 
 class CreatureFakerIT : DescribeSpec({
-    `every public function in each provider is invoked without exceptions`(CreatureFaker())
+    `every public function in each provider is invoked without exceptions`(CreaturesFaker())
 
     `faker instance is initialized with custom locale` { faker { } }
 })
