+++ faker/creatures/src/main/kotlin/io/github/serpro69/kfaker/creatures/provider/Bird.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package io.github.serpro69.kfaker.creature.provider
+package io.github.serpro69.kfaker.creatures.provider
 
 import io.github.serpro69.kfaker.FakerService
 import io.github.serpro69.kfaker.dictionary.Category
