+++ faker/creatures/src/main/kotlin/io/github/serpro69/kfaker/creatures/provider/Dog.kt
@@ -1,4 +1,4 @@
-package io.github.serpro69.kfaker.creature.provider
+package io.github.serpro69.kfaker.creatures.provider
 
 import io.github.serpro69.kfaker.FakerService
 import io.github.serpro69.kfaker.dictionary.Category
@@ -12,10 +12,10 @@ import io.github.serpro69.kfaker.provider.unique.UniqueProviderDelegate
  * [FakeDataProvider] implementation for [YamlCategory.CREATURE] category.
  */
 @Suppress("unused")
-class Horse internal constructor(fakerService: FakerService) : YamlFakeDataProvider<Horse>(fakerService) {
+class Dog internal constructor(fakerService: FakerService) : YamlFakeDataProvider<Dog>(fakerService) {
     override val yamlCategory = YamlCategory.CREATURE
-    override val secondaryCategory: Category = Category.ofName("HORSE")
-    override val localUniqueDataProvider = LocalUniqueDataProvider<Horse>()
+    override val secondaryCategory: Category = Category.ofName("DOG")
+    override val localUniqueDataProvider = LocalUniqueDataProvider<Dog>()
     override val unique by UniqueProviderDelegate(localUniqueDataProvider)
 
     init {
@@ -24,4 +24,9 @@ class Horse internal constructor(fakerService: FakerService) : YamlFakeDataProvi
 
     fun name() = resolve(secondaryCategory, "name")
     fun breed() = resolve(secondaryCategory, "breed")
+    fun sound() = resolve(secondaryCategory, "sound")
+    fun memePhrase() = resolve(secondaryCategory, "meme_phrase")
+    fun age() = resolve(secondaryCategory, "age")
+    fun coatLength() = resolve(secondaryCategory, "coat_length")
+    fun size() = resolve(secondaryCategory, "size")
 }
